fix(migrations): replace deprecated createTableIfNotExists

+++ knex/migrations/20200723144756_level_theme.ts
@@ -4,23 +4,26 @@ import { setUTF8Table } from "../utils/set-utf8-table.migration";
 
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.createTableIfNotExists(LEVEL_THEME_TABLE, (table) => {
-        setUTF8Table(table);
-        table.increments('id');
-        table.integer('order', 3).defaultTo(0);
+    const hasTable = await knex.schema.hasTable(LEVEL_THEME_TABLE);
+    if (!hasTable) {
+        await knex.schema.createTable(LEVEL_THEME_TABLE, (table) => {
+            setUTF8Table(table);
+            table.increments('id');
+            table.integer('order', 3).defaultTo(0);
 
-        table.integer('themeId').unsigned().notNullable();
-        table.integer('levelId').unsigned().notNullable();
+            table.integer('themeId').unsigned().notNullable();
+            table.integer('levelId').unsigned().notNullable();
 
-        table.foreign('themeId').references('theme.id').onDelete('CASCADE');
-        table.foreign('levelId').references('level.id').onDelete('CASCADE');
+            table.foreign('themeId').references('theme.id').onDelete('CASCADE');
+            table.foreign('levelId').references('level.id').onDelete('CASCADE');
 
-        table.index('order');
-        table.index('themeId');
-        table.index('levelId');
+            table.index('order');
+            table.index('themeId');
+            table.index('levelId');
 
-        table.unique(['themeId', 'levelId']);
-    })
+            table.unique(['themeId', 'levelId']);
+        })
+    }
 }
