+++ knex/migrations/20200814112246_comment_table.ts
@@ -5,18 +5,21 @@ import { USER_TABLE } from "../../src/entities/user.entity";
 
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.createTable(COMMENT_TABLE, (table) => {
-        setUTF8Table(table);
-        table.increments('id').unsigned();
-        table.string('text', 2000).notNullable();
-        table.integer('userId').unsigned().notNullable().references(`${USER_TABLE}.id`).onDelete('CASCADE');
-        table.integer('parentCommentId').unsigned().references(`${COMMENT_TABLE}.id`).onDelete('CASCADE');
-        table.timestamp('createdAt').notNullable().defaultTo(knex.fn.now());
+    const hasTable = await knex.schema.hasTable(COMMENT_TABLE);
+    if (!hasTable) {
+        await knex.schema.createTable(COMMENT_TABLE, (table) => {
+            setUTF8Table(table);
+            table.increments('id').unsigned();
+            table.string('text', 2000).notNullable();
+            table.integer('userId').unsigned().notNullable().references(`${USER_TABLE}.id`).onDelete('CASCADE');
+            table.integer('parentCommentId').unsigned().references(`${COMMENT_TABLE}.id`).onDelete('CASCADE');
+            table.timestamp('createdAt').notNullable().defaultTo(knex.fn.now());
 
-        table.index('createdAt');
-        table.index('userId');
-        table.index('parentCommentId');
-    });
+            table.index('createdAt');
+            table.index('userId');
+            table.index('parentCommentId');
+        });
+    }
 }
